Revert "Only build daligner using static libraries"
This reverts commit 3d7245dc0c59fd4bdeade51fd17fa938edda727f.

I had complained about using a static library for dazzdb, but
instead it was dropped for daligner, which did not need a
library in the first place.

If shared libraries are now working, there is no
reason not to use them. Still, that commit was helpful to show
how to use static libs if we ever want them.